ref: share copyLocationFrom across Lang value types via a trait
